+++ tests/e2e/modules/clock_spec.js
@@ -85,6 +85,19 @@ describe("Clock module", function () {
 		});
 	});
 
+	describe("with showTime config disabled", function () {
+		beforeAll(function () {
+			// Set config sample for use in test
+			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/clock_showTime.js";
+		});
+
+		it("should show not show the time when digital clock is shown", async function () {
+			await app.client.waitUntilWindowLoaded();
+			const time = await app.client.$$(".clock .digital .time");
+			return expect(time.length).toBe(0);
+		});
+	});
+
 	describe("with showWeek config enabled", function () {
 		beforeAll(function () {
 			// Set config sample for use in test